backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
PyPDF2==3.0.1
langchain==0.1.0
langchain-google-genai==0.0.6
pinecone-client==2.2.4
google-generativeai==0.3.2
httpx==0.25.2